requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
anyio==4.11.0
    # via watchfiles
beautifulsoup4==4.14.2
    # via
    #   pelican-image-process
    #   seafoam
blinker==1.9.0
    # via pelican
build==1.3.0
    # via pip-tools
click==8.3.0
    # via pip-tools
colorama==0.4.6
    # via
    #   build
    #   click
docutils==0.22.2
    # via pelican
feedgenerator==2.2.1
    # via pelican
html5lib==1.1
    # via -r requirements.in
idna==3.10
    # via anyio
invoke==2.2.0
    # via -r requirements.in
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   minchin-pelican-plugins-cname
    #   pelican
lxml==6.0.2
    # via pelican-image-process
markdown==3.9
    # via -r requirements.in
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
minchin-pelican-plugins-autoloader==1.2.1
    # via
    #   -r requirements.in
    #   minchin-pelican-plugins-cname
    #   minchin-pelican-plugins-nojekyll
    #   minchin-pelican-plugins-optimize-images
minchin-pelican-plugins-cname==1.3.4
    # via -r requirements.in
minchin-pelican-plugins-nojekyll==1.2.0
    # via -r requirements.in
minchin-pelican-plugins-optimize-images==1.2.2
    # via -r requirements.in
ordered-set==4.1.0
    # via pelican
packaging==25.0
    # via build
pelican==4.11.0
    # via
    #   -r requirements.in
    #   minchin-pelican-plugins-autoloader
    #   minchin-pelican-plugins-cname
    #   minchin-pelican-plugins-nojekyll
    #   minchin-pelican-plugins-optimize-images
    #   pelican-image-process
    #   pelican-jinja-filters
    #   seafoam
pelican-extended-sitemap==1.2.3
    # via -r requirements.in
pelican-image-process==3.1.0
    # via seafoam
pelican-jinja-filters==2.1.1
    # via seafoam
pillow==11.3.0
    # via
    #   -r requirements.in
    #   pelican-image-process
pip-tools==7.5.1
    # via -r requirements.in
pygments==2.18.0
    # via
    #   minchin-pelican-plugins-cname
    #   pelican
    #   rich
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via pelican
rich==14.1.0
    # via pelican
seafoam==2.13.0
    # via -r requirements.in
semantic-version==2.10.0
    # via
    #   minchin-pelican-plugins-autoloader
    #   seafoam
six==1.17.0
    # via
    #   html5lib
    #   python-dateutil
smartypants==2.0.2
    # via typogrify
sniffio==1.3.1
    # via anyio
soupsieve==2.8
    # via beautifulsoup4
titlecase==2.4.1
    # via pelican-jinja-filters
typing-extensions==4.15.0
    # via beautifulsoup4
typogrify==2.1.0
    # via -r requirements.in
tzdata==2025.2
    # via
    #   feedgenerator
    #   pelican
unidecode==1.4.0
    # via pelican
watchfiles==1.1.0
    # via pelican
webencodings==0.5.1
    # via html5lib
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
